tetragon: allow negative override return values
The user specified override value is stored as an int32 and then cast to
uint64. When the override value is negative, this casting creates a very
large number. Later when we parse this large constant, we try to store
it in a signed integer, which cannot accommodate a value this large.

Hence, write the override return value as a signed value when auto
generating the register overrides. This way, ParseAssignment will be
able to parse the negative integers.

+++ pkg/selectors/kernel.go
@@ -1099,7 +1099,7 @@ func ParseMatchAction(k *KernelSelectorState, action *v1alpha1.ActionSelector, a
 		WriteSelectorUint32(&k.data, action.ArgName)
 	case ActionTypeOverride:
 		if k.isUprobe {
-			id, err := parseOverrideRegs(k, action.ArgRegs, uint64(action.ArgError))
+			id, err := parseOverrideRegs(k, action.ArgRegs, action.ArgError)
 			if err != nil {
 				return err
 			}
